Given a parser whose hot part looks like this:

```scala
    val fieldName = Rfc7230.token.string
    val fieldValue = P.repSep0(Rfc5234.vchar.rep, P.charIn(" \t").rep).string.surroundedBy(Rfc7230.ows)
    val header = ((fieldName <* P.char(':')) ~ fieldValue).map {
      case (k, v) => Header.Raw(CIString(k), v)
    }
    val headers = (header <* P.string("\r\n")).rep0
```

`CharIn.makeError` is dominating the stack trace and allocations.  It is created at the end of each repetition, only to be nulled out and gc'ed if we've made any progress.

```
[info] Benchmark                                                         Mode  Cnt       Score       Error   Units
[info] Http1DecoderBench.http1Decoder                                   thrpt    5  194717.904 ± 27142.805   ops/s
[info] Http1DecoderBench.http1Decoder:·gc.alloc.rate                    thrpt    5    2375.264 ±   330.882  MB/sec
[info] Http1DecoderBench.http1Decoder:·gc.alloc.rate.norm               thrpt    5   13436.503 ±     0.524    B/op
[info] Http1DecoderBench.http1Decoder:·gc.churn.G1_Eden_Space           thrpt    5    2378.683 ±   323.143  MB/sec
[info] Http1DecoderBench.http1Decoder:·gc.churn.G1_Eden_Space.norm      thrpt    5   13456.252 ±   205.497    B/op
[info] Http1DecoderBench.http1Decoder:·gc.churn.G1_Survivor_Space       thrpt    5       0.008 ±     0.004  MB/sec
[info] Http1DecoderBench.http1Decoder:·gc.churn.G1_Survivor_Space.norm  thrpt    5       0.044 ±     0.019    B/op
[info] Http1DecoderBench.http1Decoder:·gc.count                         thrpt    5     508.000              counts
[info] Http1DecoderBench.http1Decoder:·gc.time                          thrpt    5     552.000                  ms
[info] Http1DecoderBench.http1Decoder:·stack                            thrpt              NaN                 ---
```

Here, we wrap every `Chain[Expectation]` in an `Eval`, so that the expectations can be created lazily.  Care must be taken not to close over mutable state, namely, `state.offset`.  We significantly increase speed and reduce garbage:

```
[info] Http1DecoderBench.http1Decoder                                   thrpt    5  353576.391 ± 39077.640   ops/s
[info] Http1DecoderBench.http1Decoder:·gc.alloc.rate                    thrpt    5    2293.983 ±   253.014  MB/sec
[info] Http1DecoderBench.http1Decoder:·gc.alloc.rate.norm               thrpt    5    7146.434 ±     0.264    B/op
[info] Http1DecoderBench.http1Decoder:·gc.churn.G1_Eden_Space           thrpt    5    2294.376 ±   247.756  MB/sec
[info] Http1DecoderBench.http1Decoder:·gc.churn.G1_Eden_Space.norm      thrpt    5    7147.791 ±    94.745    B/op
[info] Http1DecoderBench.http1Decoder:·gc.churn.G1_Survivor_Space       thrpt    5       0.007 ±     0.002  MB/sec
[info] Http1DecoderBench.http1Decoder:·gc.churn.G1_Survivor_Space.norm  thrpt    5       0.023 ±     0.006    B/op
[info] Http1DecoderBench.http1Decoder:·gc.count                         thrpt    5     494.000              counts
[info] Http1DecoderBench.http1Decoder:·gc.time                          thrpt    5     519.000                  ms
[info] Http1DecoderBench.http1Decoder:·stack                            thrpt              NaN                 ---
```

I'll work on cleaning up and providing a runnable benchmark, but wanted to get early feedback.

We might also make it leaner with simple thunks rather than an `Eval` that's always `later`.