@rossabaker 2021-10-15
Drat.  I was afraid stack safety would be an issue with the thunks.  I'll pause here and see if people agree with the problem before I pour more into the solution.