@rossabaker 2021-10-15
A thunk is slightly lighter than `Eval`, but not much more so, and Eval does indeed take care of stack safety.

[Replacing `Eval.later` with `Eval.always`](https://github.com/rossabaker/cats-parse/commit/098d02000d427f4dfa62578bbfdf3828deb58f89) is marginally slower, but eliminates some allocations:

```
[info] Http1DecoderBench.http1Decoder                                   thrpt    5  337942.154 ± 42747.443   ops/s
[info] Http1DecoderBench.http1Decoder:·gc.alloc.rate                    thrpt    5    2072.398 ±   261.164  MB/sec
[info] Http1DecoderBench.http1Decoder:·gc.alloc.rate.norm               thrpt    5    6754.582 ±     0.565    B/op
[info] Http1DecoderBench.http1Decoder:·gc.churn.G1_Eden_Space           thrpt    5    2075.145 ±   264.493  MB/sec
[info] Http1DecoderBench.http1Decoder:·gc.churn.G1_Eden_Space.norm      thrpt    5    6763.527 ±   113.341    B/op
[info] Http1DecoderBench.http1Decoder:·gc.churn.G1_Survivor_Space       thrpt    5       0.008 ±     0.003  MB/sec
[info] Http1DecoderBench.http1Decoder:·gc.churn.G1_Survivor_Space.norm  thrpt    5       0.025 ±     0.009    B/op
[info] Http1DecoderBench.http1Decoder:·gc.count                         thrpt    5     480.000              counts
[info] Http1DecoderBench.http1Decoder:·gc.time                          thrpt    5     520.000                  ms
[info] Http1DecoderBench.http1Decoder:·stack                            thrpt              NaN                 ---
```