@rossabaker 2021-10-15
I need my CPU back so I can't run a proper number of iterations right now, but looking good on the JSON:

`bench/jmh:run -i 4 -wi 2 -f 1 -t 1 .*catsParseParse`

Before, on 9e67814:

```
[info] Benchmark                   Mode  Cnt   Score    Error  Units
[info] BarBench.catsParseParse     avgt    4  ≈ 10⁻⁴           ms/op
[info] Bla25Bench.catsParseParse   avgt    4  36.105 ±  4.108  ms/op
[info] Qux2Bench.catsParseParse    avgt    4  10.878 ±  1.018  ms/op
[info] Ugh10kBench.catsParseParse  avgt    4  84.775 ±  4.723  ms/op
```

After, on 9561249:

```
[info] Benchmark                   Mode  Cnt   Score    Error  Units
[info] BarBench.catsParseParse     avgt    4  ≈ 10⁻⁴           ms/op
[info] Bla25Bench.catsParseParse   avgt    4  30.502 ±  1.736  ms/op
[info] Qux2Bench.catsParseParse    avgt    4   9.649 ±  1.508  ms/op
[info] Ugh10kBench.catsParseParse  avgt    4  76.212 ±  6.961  ms/op
```

Error on both:

```
[info] java.lang.RuntimeException: Error(256941,NonEmptyList(InRange(256941,,,,)))
[info] 	at scala.sys.package$.error(package.scala:27)
[info] 	at cats.parse.bench.JmhBenchmarks.catsParseParse(JsonBench.scala:47)
[info] 	at cats.parse.bench.jmh_generated.CountriesBench_catsParseParse_jmhTest.catsParseParse_avgt_jmhStub(CountriesBench_catsParseParse_jmhTest.java:190)
[info] 	at cats.parse.bench.jmh_generated.CountriesBench_catsParseParse_jmhTest.catsParseParse_AverageTime(CountriesBench_catsParseParse_jmhTest.java:152)
```